worker manager works, all ui is ready. there is an issue with rows that are not getting updated

+++ app/src/main/java/com/example/postpc_ex8/RootsFinder.java
@@ -5,10 +5,10 @@ import java.util.UUID;
 public class RootsFinder {
 
     UUID id;
-    int number;
+    Long number;
     String preffix;
     String suffix;
-    int progress; // todo: it needs to hold the iterNumber/NumberToCalculate as the progress of the worker??
+    Long progress; // todo: it needs to hold the iterNumber/NumberToCalculate as the progress of the worker??
     // todo: hold a worker too??
 
     public RootsFinder()
@@ -16,14 +16,14 @@ public class RootsFinder {
         // todo: is it necessary? because of serelaizblity???
     }
 
-    public RootsFinder(int num,String pre, String suff, int progressState)
+    public RootsFinder(Long num,String pre, String suff, Long progressState)
     {
         number = num;
         preffix = pre;
         suffix = suff;
         progress = progressState; //todo: maybe update this from worker?
     }
-    public RootsFinder(int num,String pre, String suff, int progressState,UUID finderId)
+    public RootsFinder(Long num,String pre, String suff, Long progressState,UUID finderId)
     {
         number = num;
         preffix = pre;
@@ -34,11 +34,11 @@ public class RootsFinder {
 
     public String getPreffix() {return this.preffix;}
     public String getSuffix() {return this.suffix;}
-    public int getProgress() {return this.progress;}
-    public int getNumber() {return this.number;}
+    public Long getProgress() {return this.progress;}
+    public Long getNumber() {return this.number;}
     public UUID getId(){return this.id;}
     public void setId(UUID finderId) {this.id = finderId;}
-    public void setProgress(int progress){this.progress = progress;}
+    public void setProgress(Long progress){this.progress = progress;}
     public void setPreffix(String newPreffix){this.preffix = newPreffix;}
     public void setSuffix(String newSuffix){this.suffix = newSuffix;}
